Move java-related code to the azure-intellij-plugin-storage-java library
