+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-storage-java/build.gradle
@@ -3,8 +3,15 @@ intellij {
 }
 
 dependencies {
+    compileOnly project(path: ':azure-intellij-plugin-lib')
+    runtimeOnly project(path: ':azure-intellij-plugin-lib', configuration: 'instrumentedJar')
     compileOnly project(path: ':azure-intellij-resource-connector-lib-java')
     runtimeOnly project(path: ':azure-intellij-resource-connector-lib-java', configuration: 'instrumentedJar')
+    compileOnly project(path: ':azure-intellij-resource-connector-lib')
+    runtimeOnly project(path: ':azure-intellij-resource-connector-lib', configuration: 'instrumentedJar')
     compileOnly project(path: ':azure-intellij-plugin-storage')
     runtimeOnly project(path: ':azure-intellij-plugin-storage', configuration: 'instrumentedJar')
+    implementation 'com.microsoft.azure:azure-toolkit-storage-lib'
+    implementation 'com.microsoft.azure:azure-toolkit-ide-common-lib'
+    implementation 'com.microsoft.azure:azure-toolkit-ide-storage-lib'
 }
