+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-storage-java/src/main/java/com/microsoft/azure/toolkit/intellij/storage/code/function/FunctionStorageAccountResourceReferenceContributor.java
